move axle.game.lazyChain to axle.algebra
